Media Queries(mobile s,m,l):Done Currently working media query for each of the screen sizes. Done with the media query for mobile small, medium, and large.
Notes/reflection:
Javascript can overwrite css. Today, I was running into a problem in which
navlinks were not showing as the css property for the nav bar was still at
display:none even though I made it display:block. But I fixed it by adding
an additional if statement to the myFunction() and made it match the media
query by making it work if the width is less that 568px.

Tmrw task:
media queries for Tablet 768px